Move SmartHint converters into their own files
